added Welcome.js, refactored page load API call
